Updated version to rev 04 in README and installer.
Added a finish page to the installer so that the user can optionally view the README upon installation.

Fixed issue in Random where saving while paused caused it to reload in such a way that the sounds were in a Halt state and unable to be Resumed, causing a failure to unpause.

Modified Random section to use an optional new mode to determine period as a multiple of the base sound length rather than in seconds. Ensured reverse compatibility.